pyproject.toml
[project]
name = "nextgen-agentic-finance"
version = "0.1.0"
description = "Agentic Finance Bootcamp – local project"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "agno>=2.1.4",
    "chonkie[st]>=1.4.0",
    "ddgs>=9.6.1",
    "duckdb>=1.4.1",
    "fastapi>=0.119.0",
    "huggingface-hub>=0.35.3",
    "lancedb>=0.25.2",
    "mcp[cli]>=1.19.0",
    "openai>=2.3.0",
    "pandas>=2.3.3",
    "pypdf>=6.1.1",
    "quantmod==0.1.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "youtube-transcript-api>=1.2.3",
    "python-dotenv>=1.0",
    "requests>=2.32",
    "pytest>=8.0",
    "matplotlib>=3.9"
]
